refactor: ticket payment component relocated to the components file
