+++ src/components/TicketPayment/styles/CardSectionStyle.js
