+++ src/components/TicketPayment/styles/CommomStyle.js
