+++ src/pages/Dashboard/Payment/index.js
@@ -2,8 +2,8 @@ import Splash from '../../../components/Splash';
 import useTicket from '../../../hooks/api/useTicket';
 import useEnrollment from '../../../hooks/api/useEnrollment';
 import NoTicket from './NoTicket';
-import HasReservedTicket from './HasReservedTicket';
 import NoEnrollment from './NoEnrollment';
+import HasReservedTicket from '../../../components/TicketPayment';
 
 export default function Payment() {
   const { ticket, ticketLoading, getTicket } = useTicket();
